+++ texecom_alarm/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2] - 2026-08-24
+
 ### Added
 
 - Ready-to-arm switches (Away, Home, Night) that start on. When a switch is